Move: 프로필 디렉토리 분류 작업
