+++ src/pages/Profile/ProfilePage.tsx
@@ -1,4 +1,4 @@
-import Profile from "../../components/Profile/Profile";
+import Profile from "../../components/Profile/Main/Profile";
 
 function ProfilePage(): JSX.Element {
   return (
